third_party/WebKit/LayoutTests/svg/custom/disallow-non-lengths-in-attrs-expected.txt

CONSOLE ERROR: line 22: Error: <svg> attribute width: Expected length, "auto".
CONSOLE ERROR: line 22: Error: <foreignObject> attribute width: Expected length, "auto".
CONSOLE ERROR: line 22: Error: <svg> attribute width: Expected length, "initial".
CONSOLE ERROR: line 22: Error: <foreignObject> attribute width: Expected length, "initial".
CONSOLE ERROR: line 22: Error: <svg> attribute width: Expected length, "inherit".
CONSOLE ERROR: line 22: Error: <foreignObject> attribute width: Expected length, "inherit".
CONSOLE ERROR: line 22: Error: <svg> attribute width: Expected length, "foo".
CONSOLE ERROR: line 22: Error: <foreignObject> attribute width: Expected length, "foo".
This is a testharness.js-based test.
PASS Test width 'auto' on SVGSVGElement
PASS Test width 'auto' on SVGForeignObject
PASS Test width 'initial' on SVGSVGElement
PASS Test width 'initial' on SVGForeignObject
PASS Test width 'inherit' on SVGSVGElement
PASS Test width 'inherit' on SVGForeignObject
PASS Test width 'foo' on SVGSVGElement
PASS Test width 'foo' on SVGForeignObject
Harness: the test ran to completion.
